+++ DlgProcSetupPen_Color.cpp
@@ -1,5 +1,12 @@
 #include "stdafx.h"
 
+#include <Windows.h>
+
+#include <afx.h>
+#include <afxwin.h>
+
+#include <atltypes.h>
+
 #include <cstdlib>
 
 #include "PegAEsys.h"
@@ -24,7 +31,7 @@ BOOL CALLBACK DlgProcSetupPenColor(HWND hDlg, UINT anMsg, WPARAM wParam, LPARAM 
     case WM_INITDIALOG:
         for (size_t i = 0; i < sizeof(crHotCols) / sizeof(COLORREF); i++)
         {
-            _itoa_s(i, szBuf, sizeof(szBuf), 10);
+            _itoa_s(static_cast<int>(i), szBuf, sizeof(szBuf), 10);
             ::SendDlgItemMessage(hDlg, IDC_COL_LIST, CB_ADDSTRING, 0, (LPARAM)(LPCSTR)szBuf);
         }
         nPenColor = pstate.PenColor();
